+++ src/main/java/teamkim/stream/domain/logging/enums/ClassType.java
@@ -1,4 +1,4 @@
-package teamkim.stream.domain.logging;
+package teamkim.stream.domain.logging.enums;
 
 public enum ClassType {
     PAPER_BOX, TRAFFIC_SIGN, CONTAINER, WOODEN_BUILDING, TREE, PLASTIC, DISPOSABLE_CUP
