+++ src/main/java/teamkim/stream/domain/logging/enums/DirectionType.java
@@ -1,4 +1,4 @@
-package teamkim.stream.domain.logging;
+package teamkim.stream.domain.logging.enums;
 
 public enum DirectionType {
     FRONT, BACK, LEFT, RIGHT
