+++ src/main/java/teamkim/stream/domain/logging/repository/LoggingRepository.java
@@ -1,6 +1,9 @@
-package teamkim.stream.domain.logging;
+package teamkim.stream.domain.logging.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import teamkim.stream.domain.logging.enums.ClassType;
+import teamkim.stream.domain.logging.enums.DirectionType;
+import teamkim.stream.domain.logging.entity.LoggingEntity;
 
 import java.util.List;
